+++ app/Core/Middleware/AuthMiddleware.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App\Core\Middleware;
+
+use App\Core\Auth;
+use App\Core\Contracts\Middleware;
+
+class AuthMiddleware implements Middleware
+{
+    public function handle()
+    {
+        if (Auth::user() == null) {
+            header("location: " . LOGIN_PAGE);
+            exit();
+        }
+    }
+}