+++ app/controllers/AboutController.php
@@ -2,6 +2,8 @@
 
 namespace App\Controllers;
 
+use App\Core\Middleware\Middleware;
+
 class AboutController extends Controller
 {
     public function index()