+++ app/controllers/PostController.php
@@ -2,6 +2,7 @@
 
 namespace App\Controllers;
 
+use App\Core\Auth;
 use App\Core\Str;
 use App\Core\Validator;
 use App\Core\FileSystem;
@@ -29,6 +30,12 @@ class PostController extends Controller
     }
     public function store()
     {
+        $user = Auth::user();
+
+        if ($user == null) {
+            // TODO: UPDATE to 401
+            abort(404);
+        }
         extract($this->data);
         $data = $_POST;
 
@@ -60,14 +67,14 @@ class PostController extends Controller
             $data['description'] = trim(htmlspecialchars($data['description']));
             $data['body'] = trim(htmlspecialchars($data['body']));
             $data["slug"] = trim(Str::toKebabCase($data['title']));
-            $data["author_id"] = 1;
+            $data["author_id"] = $user['id'];
             $data["module_id"] = null;
 
             $thumbnail_file = null;
 
             try {
-                $thumbnail_file = FileSystem::uploadImage($_FILES['thumbnail'], Str::randomString($data['slug']));
-                $data['thumbnail'] = $thumbnail_file;
+                $thumbnail_file = FileSystem::uploadImage($_FILES['thumbnail']);
+                $data['thumbnail'] = '/' . $thumbnail_file;
                 if (!$validator->file_path($thumbnail_file) || $thumbnail_file == null) {
                     $errors['thumbnail'] = "Thumbnail was not uploaded successfully!";
                 }
@@ -80,10 +87,12 @@ class PostController extends Controller
         if (empty($errors)) {
 
             PostsRepository::store($data);
+            $slug = $data["slug"];
             $data = [];
             $errors = [];
             unset($_POST);
-            header("Location: /posts");
+            header("Location: /authors/posts/" . $slug);
+            exit();
         }
 
         view("create-post", [
@@ -102,6 +111,13 @@ class PostController extends Controller
 
     public function edit()
     {
+        $user = Auth::user();
+
+        if ($user == null) {
+            // TODO: UPDATE to 401
+            abort(404);
+        }
+
         extract($this->data);
         $data = $_POST;
 
@@ -137,15 +153,15 @@ class PostController extends Controller
             $data['description'] = trim(htmlspecialchars($data['description']));
             $data['body'] = trim(htmlspecialchars($data['body']));
             $data["slug"] = trim(Str::toKebabCase($data['title']));
-            $data["author_id"] = 1;
+            $data["author_id"] = $user['id'];
             $data["module_id"] = null;
 
             if ($data['thumbnail'] != null) {
                 $thumbnail_file = null;
 
                 try {
-                    $thumbnail_file = FileSystem::uploadImage($_FILES['thumbnail'], Str::randomString($data['slug']));
-                    $data['thumbnail'] = $thumbnail_file;
+                    $thumbnail_file = FileSystem::uploadImage($_FILES['thumbnail']);
+                    $data['thumbnail'] = '/' . $thumbnail_file;
                     if (!$validator->file_path($thumbnail_file) || $thumbnail_file == null) {
                         $errors['thumbnail'] = "Thumbnail was not uploaded successfully!";
                     }
@@ -158,10 +174,12 @@ class PostController extends Controller
 
         if (empty($errors)) {
             PostsRepository::update($data);
+            $slug = $data["slug"];
             $data = [];
             $errors = [];
             unset($_POST);
-            header("Location: /posts");
+            header("Location: /authors/posts/" . $slug);
+            exit();
         }
 
         view("update-post", [
@@ -173,9 +191,20 @@ class PostController extends Controller
 
     public function destroy()
     {
+        $user = Auth::user();
+
+        if ($user == null) {
+            // TODO: UPDATE to 401
+            abort(404);
+        }
+
         try {
-            PostsRepository::delete(['id' => $_POST['id']]);
-            header("location: /posts");
+
+            PostsRepository::delete(['id' => $_POST['id'], 'author_id' => $user['id']]);
+            $routeBase = explode('/', $_SERVER['HTTP_REFERER'])[3];
+            $redirect = $routeBase === 'authors' ? 'location: /authors/posts' : 'location: /posts';
+            header($redirect);
+            exit();
         } catch (Exception $ex) {
             dd($ex->getMessage());
         }
@@ -183,12 +212,23 @@ class PostController extends Controller
 
     public function updateDraftState()
     {
+        $user = Auth::user();
+
+        if ($user == null) {
+            // TODO: UPDATE to 401
+            abort(404);
+        }
+
         try {
             PostsRepository::updateDraftState([
                 'id' => $_POST['id'],
-                'draft' => !$_POST['draft']
+                'draft' => !$_POST['draft'],
+                'author_id' => $user['id']
             ]);
-            header("location: /posts");
+            $routeBase = explode('/', $_SERVER['HTTP_REFERER'])[3];
+            $redirect = $routeBase === 'authors' ? 'location: /authors/posts' : 'location: /posts';
+            header($redirect);
+            exit();
         } catch (Exception $ex) {
             dd($ex->getMessage());
         }