+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 # Change log
 
+## 14/06/2023
+- Authors Page
+- Author Page
+- Authors Posts page
+- Authors Post page -- Can view draft posts also
+
+## 13/06/2023
+- Refactor Router
+- Validate Middlewares
+- Logout
+- Display Post Manager for post author
+- Validate Author id before Post CRUD
+- Fix Upload Thumbnails errors
+
+## 01/06/2023
+- Implement Middleware
+- Profile
+- Login
+- Guest vs Authenticated.
+- Default profile picture.
+- Validate unique attribute.
+- Save user to session.
+- Register new user.
+- Implement Service Container.
+- Use Session to author id.
+
+## 31/05/2023
+- Create a new author.
+
 ## 30/05/2023
 - Refactor Controller.
 - Update Upload Target Path.